+++ src/cqportal/__init__.py
@@ -1,4 +1,5 @@
 from .PortalBase import PortalBase
 from .Frame import Frame
 from .Ramp import Ramp
+from .RampGreebled import RampGreebled
 from .Portal import Portal