+++ changes.md
@@ -5,6 +5,8 @@
 ## 0.1.2
 * Refactored Ramp to have less convoluted parent if statements
 * Added ramp_frame.py example
+* Added RampGreebled
+  * Added Portal Greebled Ramp Example
 
 ## 0.1.1
 * As predicted fixed as issue with ramp inheriting width from frame in unexpected and exciting ways.